modules/location-processor/requirements.txt
kafka-python==2.0.2
SQLAlchemy==1.3.19
psycopg2-binary==2.8.5
GeoAlchemy2==0.8.4
shapely==1.7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.0
